API improvements
