+++ lib/neon-sled/package.json
@@ -1,6 +1,6 @@
 {
   "name": "neon-sled",
-  "version": "0.1.0",
+  "version": "0.1.2",
   "description": "",
   "main": "lib/index.js",
   "author": [
