--- lib/node/native/artifacts.json
@@ -1,1 +1,0 @@
-{"active":"release","targets":{"release":{"rustc":"rustc 1.20.0 (f3d6973f4 2017-08-27)","env":{"npm_config_target":null,"npm_config_arch":null,"npm_config_target_arch":null,"npm_config_disturl":null,"npm_config_runtime":null,"npm_config_build_from_source":null,"npm_config_devdir":null}}}}
